requirements.txt
absl-py==1.2.0
astor==0.8.1
astunparse==1.6.3
black==22.8.0
certifi==2016.2.28
click==8.0.4
cloudpickle==2.2.0
contextlib2==21.6.0
dataclasses==0.8
decorator==4.4.2
dm-sonnet==1.36
dm-tree==0.1.7
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
graph-nets==1.1.0
grpcio==1.48.2
h5py==2.10.0
importlib-metadata==4.8.3
install==1.3.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.7
mypy-extensions==0.4.3
networkx==2.5.1
numpy==1.19.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvcc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.7.2.91
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-dali-cuda110==1.16.0
nvidia-dali-nvtf-plugin==1.16.0+nv22.8
nvidia-horovod==0.25.0+nv22.8
nvidia-nccl-cu11==2.14.3
nvidia-pyindex==1.0.9
nvidia-tensorboard==1.15.0+nv21.4
nvidia-tensorflow==1.15.5+nv22.8
nvidia-tensorrt==8.4.3.1
opt-einsum==3.3.0
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.19.5
psutil==5.9.2
PyYAML==6.0
semantic-version==2.10.0
six==1.16.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-probability==0.8.0
termcolor==1.1.0
tomli==1.2.3
typed-ast==1.5.4
typing_extensions==4.1.1
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.14.1
zipp==3.6.0
